Move `AnyConsensusState` to the relayer crate
