+++ relayer/src/chain/runtime.rs
@@ -8,7 +8,6 @@ use tracing::error;
 use ibc::{
     core::{
         ics02_client::{
-            client_consensus::{AnyConsensusState, AnyConsensusStateWithHeight},
             client_state::{AnyClientState, IdentifiedAnyClientState},
             events::UpdateClient,
         },
@@ -31,6 +30,7 @@ use ibc::{
 
 use crate::{
     account::Balance,
+    client_consensus::{AnyConsensusState, AnyConsensusStateWithHeight},
     config::ChainConfig,
     connection::ConnectionMsgType,
     denom::DenomTrace,
